Switched names so ResultPage is named ResultPage. Updated all imports.
