+++ src/main.tsx
@@ -8,13 +8,11 @@ import StartPage from './pages/StartPage'
 
 const router =  createBrowserRouter(
   createRoutesFromElements(
-    <Route path="/" element={<App />}>
-      
-      {/* <Route path="result/:ings" element={<ResultPage />}></Route> 
-      <Route path="" element={<StartPage/>} /> */}
+    <Route path="/" element={<App />}><>
+      <Route path="result/:ings" element={<ResultPage />}></Route>
+      <Route path="" element={<StartPage />} /></>
       <Route path="" element={<RecipePage/>} />
       <Route path="*" element={<h2>404 not found</h2>} />
-
     </Route>
 
   )
