+++ src/pages/RecipePage.tsx
@@ -1,66 +1,78 @@
-import ListItem from "../components/ListItem"
-import RecipeComponent from "../components/RecipeCard"
-import { RecipeCard } from "../interfaces"
-import styled from "styled-components";
-import { useState } from "react";
-import bg from "../assets/bg.jpg";
+import { useEffect, useState } from "react";
+import { useParams } from "react-router-dom";
+import { RecipeCard } from "../interfaces";
 
+
+
+import styled from 'styled-components';
+
+function ResultPage() {
+    const {ings} = useParams();
+    const [ingredients, setIngredients] = useState(ings);
+    const [recipeCards, setRecipeCards] = useState<RecipeCard[]>();
+
+    /**
+     * Calls API on change of the ingredients state.
+     */
+    useEffect(() => {
+        const options = {
+            // API key: ea186f9a58784d0d86b47956204c76be
+            method: "GET",
+            headers: {
+              "Content-Type": "application/json",
+            },
+          }
+
+        fetch(`https://api.spoonacular.com/recipes/findByIngredients?apiKey=ea186f9a58784d0d86b47956204c76be&ingredients=${ingredients}&number=3`,
+        options)
+        .then((r) => r.json())
+        .then((data) => setRecipeCards(data));
+    }, [ingredients] 
+    )
+
+    /**
+     * Removes hyphens from ingredients params and replaces with spaces.
+     * Separates ingredients by ',+'.
+     * Return array of li elements with ingrdient names as content and keys.
+     */
+    function listIngredients() {
+        if(ings)
+        return ings.replace("-", " ").split(",+").map(ing => <li key={ing}>{ing}</li>);
+    }
+
+    /**
+     * Returns array of list items for each recipe in the state.
+     */
+    function listRecipes() {
+        if (recipeCards)
+        return recipeCards.map((r) => {return <li key={r.title}>{r.title}</li>})
+    }
+
+    return(
+        <PageContainer>
+            <div>
+                <ul>
+                    {listIngredients()}
+                </ul>
+            </div>
+            <div>
+                <ul>
+                    {recipeCards ? listRecipes() : null}
+                </ul>
+            </div>
+        </PageContainer>
+    )
+}
 
 const PageContainer = styled.div`
   display: flex;
   flex-direction: column;
+  justify-content: center;
   align-items: center;
-  background-color: #141414;
+  background-color: #7a949c;
   color: #efefef;
+  padding: 1.2rem;
   min-height: 100vh;
-  background-image: url(${bg});
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
 `;
 
-const IngredientsContainer = styled.div`
-    width: 100%;
-    max-width: 500px;
-    display: flex;
-    justify-content: space-evenly;
-    flex-direction: row;
-    flex-wrap: wrap;
-    margin: 0;
-    padding-top: 1.5rem;
-`;
-
-const RecipesContainer = styled.div`
-    overflow-y: scroll;
-    max-height: 600px;
-`
-
-function RecipePage() {
-
-
- return(
-    <PageContainer>
-        <IngredientsContainer>
-            <ListItem text={""} onDelete={function (): void {
-                throw new Error("Function not implemented.")
-            } }></ListItem>
-            <ListItem text={""} onDelete={function (): void {
-                throw new Error("Function not implemented.")
-            } }></ListItem>
-            <ListItem text={""} onDelete={function (): void {
-                throw new Error("Function not implemented.")
-            } }></ListItem>
-        </IngredientsContainer>
-        
-        <RecipesContainer>
-            <RecipeComponent></RecipeComponent>
-            <RecipeComponent></RecipeComponent>
-            <RecipeComponent></RecipeComponent>
-            <RecipeComponent></RecipeComponent>
-        </RecipesContainer>
-    </PageContainer>
-)
-}
-
-
-export default RecipePage
+export default ResultPage
